# public/openapi.yaml
...
openapi: 3.0.3
info:
  title: AccessGate API
  version: 0.1.0
  description: Event management and ticketing system API
  contact:
    name: API Support
    email: azeezwaris17@gmail.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api/v1
    description: Development server
...
paths:
...
  /events:
    get:
      tags:
        - Events
      summary: Get all events with filtering and pagination
      description: Retrieve all events for the authenticated admin with optional filtering and pagination
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [all, active, completed, cancelled]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            default: 10
      responses:
        '200':
          description: Successfully retrieved events
          content:
            application/json:
              example:
                success: true
                events: []
                pagination:
                  currentPage: 1
                  totalPages: 1
                  totalEvents: 0
                  hasNext: false
                  hasPrev: false
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
...
  /tickets/validate-ticket-by-code:
    post:
      tags:
        - Public
      summary: Validate ticket by code
      description: Validate a ticket using its code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ticketCode:
                  type: string
            example:
              ticketCode: "ABCD-1234-EFGH-5678"
      responses:
        '200':
          description: Successfully validated ticket
          content:
            application/json:
              example:
                ticket:
                  id: "507f1f77bcf86cd799439015"
                  ticketCode: "ABCD-1234-EFGH-5678"
                  attendeeName: "John Doe"
                  attendeeEmail: "john@example.com"
                  status: "valid"
                  event:
                    _id: "507f1f77bcf86cd799439013"
                    name: "Tech Conference 2024"
        '400':
          description: Ticket code required or ticket revoked
        '404':
          description: Ticket not found
        '500':
          description: Internal server error
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...